Initial definition of replay.proto
Game replays are represented in GameReplay (protocol buffer)[proto-doc] message which contains only the subset of fields relevant to recreating the moves of the game, personally identifiable information is not stored here (though with a full index of the game data from OSN, one could probably narrow down the candidate gcIDs from the gameplay turns).  Some rank movement is stored in the GameMatch message for determining reward/penalty of turn action selection, but in the interest of being able to group by play action (where eventual game outcome may vary) the rank movement is not made part of the GameReplay message.  The GameStatistics message represents aggregate statistics for games with a common prefix or full representation.

[proto-doc][https://developers.google.com/protocol-buffers/]